[feat #163451769]Restructure Application
