+++ server/server.js
@@ -4,7 +4,7 @@ import router from './routes/v1/index';
 
 
 const app = express();
-const port = process.env.PORT || 8000;
+const port = process.env.PORT || 8080;
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
